refactor(eve): flatten the driver loop
The driver ran a labeled outer loop with a nested unbounded wait loop,
continue-to-label on cancellation, and break-to-outer on delivery. The
cancel re-entry was only safe by accident: it re-processed the previous
park's settled answer and avoided double-notifying solely because the
caller had already been cleared.

One loop now. The wait-route cycle moves to parked-delivery-wait.ts as
nextTurnDelivery, which returns a discriminated instruction (expired,
closed, cancel-turn, turn); the driver switches on it. The cancel path
re-enters with settled explicitly cleared, so the no-double-notify
property holds by construction. The module split also keeps
workflow-entry under the 700-line structural cap.
